+++ realSemiProject/src/main/webapp/WEB-INF/views/board/list.jsp
@@ -280,7 +280,7 @@
 									</c:if>
 								
 								</c:otherwise>
-							</c:choose>
+							</c:choose	>
 						</td>
 						
 						<!-- 작성자 -->
